+++ lib/teiserver_web/controllers/report/general_controller.ex
@@ -1,0 +1,21 @@
+defmodule TeiserverWeb.Report.GeneralController do
+  use CentralWeb, :controller
+
+  plug(AssignPlug,
+    sidemenu_active: ["teiserver"]
+  )
+
+  plug Bodyguard.Plug.Authorize,
+    policy: Teiserver.Account.Admin,
+    action: {Phoenix.Controller, :action_name},
+    user: {Central.Account.AuthLib, :current_user}
+
+  plug(:add_breadcrumb, name: 'Teiserver', url: '/teiserver')
+  plug(:add_breadcrumb, name: 'Reports', url: '/teiserver/reports')
+
+  @spec index(Plug.Conn.t(), map) :: Plug.Conn.t()
+  def index(conn, _params) do
+    conn
+    |> render("index.html")
+  end
+end