+++ lib/teiserver_web/live/agents/index.ex
@@ -25,7 +25,7 @@ defmodule TeiserverWeb.AgentLive.Index do
       |> assign(:filters, %{})
       |> assign(:sidemenu_active, "teiserver")
       |> assign(:colours, Central.Admin.ToolLib.colours())
-      |> assign(:menu_override, Routes.ts_lobby_general_path(socket, :index))
+      |> assign(:menu_override, Routes.ts_general_general_path(socket, :index))
 
     {:ok, socket, layout: {CentralWeb.LayoutView, "blank_live.html"}}
   end