+++ lib/teiserver_web/views/admin/general_view.ex
@@ -7,10 +7,10 @@ defmodule TeiserverWeb.Admin.GeneralView do
   def icon(), do: StylingHelper.icon(:info)
 
   @spec colours(String.t()) :: {String.t(), String.t(), String.t()}
-  def colours("clients"), do: Teiserver.ClientLib.colours()
   def colours("clans"), do: Teiserver.Clans.ClanLib.colours()
   def colours("users"), do: Teiserver.Account.UserLib.colours()
   def colours("queues"), do: Teiserver.Game.QueueLib.colours()
   def colours("parties"), do: Teiserver.Game.PartyLib.colours()
+  def colours("telemetry"), do: Teiserver.Telemetry.TelemetryLib.colours()
   def colours("tools"), do: Central.Admin.ToolLib.colours()
 end