+++ lib/teiserver_web/views/report/general_view.ex
@@ -1,0 +1,13 @@
+defmodule TeiserverWeb.Report.GeneralView do
+  use TeiserverWeb, :view
+
+  @spec colours :: {String.t(), String.t(), String.t()}
+  def colours(), do: StylingHelper.colours(:primary)
+
+  @spec icon() :: String.t()
+  def icon(), do: StylingHelper.icon(:primary)
+
+  @spec colours(String.t()) :: {String.t(), String.t(), String.t()}
+  def colours("client_events"), do: Teiserver.Telemetry.ClientEventLib.colours()
+  def colours("server_metrics"), do: Teiserver.Telemetry.TelemetryDayLogLib.colours()
+end