+++ lib/teiserver_web/views/report/metric_view.ex
@@ -1,8 +1,11 @@
-defmodule TeiserverWeb.Telemetry.ClientEventView do
+defmodule TeiserverWeb.Report.MetricView do
   use TeiserverWeb, :view
 
-  def colours(), do: Teiserver.Telemetry.ClientEventLib.colours()
-  def icon(), do: Teiserver.Telemetry.ClientEventLib.icon()
+  @spec colours :: {String.t(), String.t(), String.t()}
+  def colours(), do: Teiserver.Telemetry.TelemetryDayLogLib.colours()
+
+  @spec icon() :: String.t()
+  def icon(), do: Teiserver.Telemetry.TelemetryDayLogLib.icon()
 
   # For the detail view in day metrics
   def heatmap(value, maximum, colour) do