+++ test/teiserver_web/controllers/lobby/general_controller_test.exs
@@ -9,7 +9,7 @@ defmodule TeiserverWeb.Lobby.GeneralControllerTest do
   end
 
   test "index", %{conn: conn} do
-    conn = get(conn, Routes.ts_lobby_general_path(conn, :index))
+    conn = get(conn, Routes.ts_general_general_path(conn, :index))
 
     assert html_response(conn, 200) =~ "Battles"
     assert html_response(conn, 200) =~ "Account"